Validate URLs when constructing Sarus object
This will better inform a user about invalid URLs when constructing a
new Sarus object. It will already give an error during construction
instead of the connect step. That way, we won't repeatedly try to
connect to the same URL, despite it being invalid the whole time.

Add test case in __tests__ for

- Completely invalid URL (fails to be fed into new URL())
- Invalid protocol (must be ws / wss)

We use stock JS functionality for this, so no custom URL parser.